requirements.txt
numpy
Pillow
glfw
PyOpenGL
pyrr